fix: deleted unnecessary file
